Add overall timeout of 10 seconds (#135)
*Issue #, if available:*
Currently there is no timeout set when plugin is executed as binary.

*Description of changes:*
This PR implements a 10-second overall execution time limit for the
plugin to execute signing and verification.

#### TODO: Have better timeout for GenerateEnvelope and Verify
operations.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Signed-off-by: Pritesh Bandi <priteshbandi@gmail.com>